php_travels/requirements.txt
pytest==5.3.4
selenium==3.141.0
webdriver-manager==2.3.0
allure-pytest==2.7.0

